+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ### 0.0.3
 * Changed `PartitionPosition` to `RangePosition` (without `PartitionId`)
+* Changed `ChangeFeedPosition` from `RangePosition list` to `RangePosition[]`
 * Made `ChangeFeedPositionTracker` internal
 
 ### 0.0.2